Fix the names of the tests so the script discovers them again
